Add 'checkAccentError' to it's own file so that it can be reused'
